Changed Altitude to int on AirportCtr

+++ FlightSystem/FlightAdmin/Controller/AirportCtr.cs
@@ -10,7 +10,7 @@ namespace FlightAdmin.Controller {
 
         #region Create / Update / Delete
 
-        public Airport CreateAirport(string name, string shortName, string city, string country, decimal latitude, decimal longtitude, decimal altitude, string timeZone) {
+        public Airport CreateAirport(string name, string shortName, string city, string country, decimal latitude, decimal longtitude, int altitude, string timeZone) {
             Airport airport;
 
             try {
@@ -39,7 +39,7 @@ namespace FlightAdmin.Controller {
             return airport;
         }
 
-        public Airport UpdateAirport(Airport airport, string name, string shortName, string city, string country, decimal latitude, decimal longtitude, decimal altitude, string timeZone) {
+        public Airport UpdateAirport(Airport airport, string name, string shortName, string city, string country, decimal latitude, decimal longtitude, int altitude, string timeZone) {
             try {
                 airport.Name = name;
                 airport.ShortName = shortName;
